Refactor ARM CPUID code
Split into arm32 and arm64 versions since they are mostly distinct code.

Remove the code for sniffing NEON/AES support based on iOS version models. All
64-bit Apple ARM processors support NEON/AES/etc, so we can just assume it is
there if we are building for aarch64. And Apple does not support 32-bit
applications at all anymore (starting in iOS 11), so there is no point in new
code supporting such systems.

Add support for a couple of new Linux/POSIX methods of getting the cache line
size using sysconf and different auxv fields.

Move the cache line detection code to os_utils since with the sole exception
of x86 we require OS specific support to get this information. This also
has the effect of enabling it on more platforms (particularly for Linux).
